Move all Cut and Top Cut logic and models
