+++ aesops/blueprints/tournament_blueprint.py
@@ -4,6 +4,7 @@ from data_models.players import Player
 from data_models.tournaments import Tournament
 from data_models.model_store import db
 import aesops.business_logic.players as p_logic
+import aesops.business_logic.top_cut as tc_logic
 import aesops.business_logic.tournament as t_logic
 from aesops.forms import (
     PlayerForm,
@@ -35,6 +36,7 @@ def tournament(tid):
         display_side_bias=display_side_bias,
         get_faction=get_faction,
         t_logic=t_logic,
+        tc_logic=tc_logic,
         p_logic=p_logic,
         last_concluded_round=tournament.current_round
         if t_logic.is_current_round_finished(tournament)
