+++ data_models/model_store.py
@@ -4,7 +4,6 @@ db = SQLAlchemy()
 
 from .match import Match
 from .players import Player
-from .top_cut import CutPlayer, ElimMatch
+from .top_cut import Cut, CutPlayer, ElimMatch
 from .tournaments import Tournament
 from .users import User
-from top_cut.cut import Cut
